[Chore] Remove the cycles between the export submodules
Making the imports honest exposed five cycles: `assetstudio <-> payload`,
`assetstudio <-> tasks`, `payload <-> paths`, `payload <-> types` and
`payload <-> images`. Rust allows module cycles, so they cost nothing at
compile time and everything at reading time -- no module in this
directory could be understood, moved or extracted on its own.

Each one turned out to be a handful of functions sitting in the wrong
file, not a genuine mutual dependency:

- `assetstudio_export_type_selector`, `assetstudio_type_selector_matches`,
  `normalize_assetstudio_type_name` and `asset_studio_export_type_list`
  are pure name mapping over `RegionConfig`. They were split across
  `assetstudio` and `tasks`, which `paths`, `payload` and each other then
  had to import: four cycles for four functions that depend on nothing.
  They are now `selectors`, a leaf.
- `image_format_extension` maps a config enum to a string; it moves to
  `types`, beside the enum.
- `safe_payload_bundle_path` sanitises a path; it moves to `paths`.
- The RGBA interchange form -- `NativeRgbaIr` and the three functions
  that parse and decode it -- is image work that lived in `payload`; it
  moves to `images`.

The result is a DAG. `limits`, `selectors`, `tasks` and `types` are
leaves; `paths` sits above them, then `images`, then `payload`, with
`assetstudio` and `media_postprocess` on top. A cycle check over the
`use super::` edges reports none.
