+++ src/core/export_pipeline/implementation/tests.rs
@@ -19,8 +19,7 @@ use crate::core::config::{
 use crate::core::errors::ExportPipelineError;
 
 use super::assetstudio::{
-    assetstudio_export_type_selector, assetstudio_object_mode_supported_type,
-    assetstudio_type_selector_matches, native_read_batch_size_for_assets,
+    assetstudio_object_mode_supported_type, native_read_batch_size_for_assets,
     native_read_kind_for_asset, native_skipped_unsupported_asset,
     select_native_object_readable_assets, sort_native_object_reads_for_failure_isolation,
 };
@@ -33,14 +32,16 @@ use super::media_postprocess::{
 };
 use super::paths::{
     assetstudio_fix_file_name, native_object_output_extension, native_object_output_path,
+    safe_payload_bundle_path,
 };
 use super::payload::{
     parse_payload_bundle, parse_payload_bundle_borrowed, playable_container_output_path,
-    safe_payload_bundle_path, text_asset_public_bytes_target,
-    write_assetstudio_export_manifest_entry, write_assetstudio_playable_payloads,
-    write_native_image_payload_final_files, write_native_image_payload_final_files_with_backend,
-    write_native_object_payload, write_native_payload_file,
+    text_asset_public_bytes_target, write_assetstudio_export_manifest_entry,
+    write_assetstudio_playable_payloads, write_native_image_payload_final_files,
+    write_native_image_payload_final_files_with_backend, write_native_object_payload,
+    write_native_payload_file,
 };
+use super::selectors::{assetstudio_export_type_selector, assetstudio_type_selector_matches};
 use super::tasks::{
     prepare_usm_processing_inputs, run_path_tasks, scan_all_files, usm_segment_key,
     UsmProcessingInput,
